rename satellite switch, split map controls container and switch into separate components.
