+++ frontend/src/components/map/MapTopToolbar.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+class MapTopToolbar extends Component {
+  render() {
+    return (
+      <div className='map-top-toolbar'>
+        {this.props.children}
+      </div>
+    );
+  }
+}
+export default MapTopToolbar;
